+++ R/check.R
@@ -97,15 +97,7 @@ checkFeatures <- function(features) {
       ncol(features[[i]]) > 0
     )
     hasUniqueIdColumn(features[[i]])
-    characterColsFilter <- vapply(features[[i]], is.character, logical(1))
-    nonCharacterCols <- colnames(features[[i]])[!characterColsFilter]
-    if (!isEmpty(nonCharacterCols)) {
-      warning(
-        "Detected non-character columns in features table input. ",
-        "The following columns were automatically coerced to character strings: ",
-        paste(nonCharacterCols, collapse = ", ")
-      )
-    }
+    warnIfNonCharacterCols(features[[i]])
   }
 
   return(NULL)
@@ -231,6 +223,7 @@ checkMetaFeatures <- function(metaFeatures, study = NULL) {
       nrow(metaFeatures[[i]]) > 0,
       ncol(metaFeatures[[i]]) > 0
     )
+    warnIfNonCharacterCols(metaFeatures[[i]])
   }
 
   return(NULL)